Minor refactoring for lisibility
