+++ bot.js
@@ -2,42 +2,34 @@ const Twit = require('twit');
 const fs = require('fs');
 const path = require('path');
 
-
-const config = {
-    consumer_key: process.env.CONSUMER_KEY,
-    consumer_secret: process.env.CONSUMER_SECRET,
-    access_token: process.env.ACCESS_TOKEN,
-    access_token_secret: process.env.ACCESS_TOKEN_SECRET
-};
+const config = require('./config');
 
 const T = new Twit(config);
 
-
+// Returning a random filename
 const random_from_array = images => images[Math.floor(Math.random() * images.length)];
 
 
 const upload_random_image = (images, statusText) => {
-    console.log('Fetching a random image');
+    // Fetching a random image's file name and path
     const image_path = path.join(__dirname, '/images/' + random_from_array(images));
 
-    console.log('Encoding the image');
+    // Encoding the file
     const b64content = fs.readFileSync(image_path, { encoding: 'base64' });
 
-    console.log('Uploading the image as media on twitter');
+    // Uploading the image as media on twitter
     T.post('media/upload', { media_data: b64content }, (err, data, response) => {
         if (err) {
-            console.log('ERROR:');
-            console.log(err);
+            console.log(`Error: ${err}`);
         } else {
-            console.log('Posting that media with a status');
+            // Posting that media with a status
             T.post('statuses/update', {
                 status: statusText,
                 media_ids: new Array(data.media_id_string)
             },
                 (err, data, response) => {
                     if (err) {
-                        console.log('Error:');
-                        console.log(err);
+                        console.log(`Error: ${err}`);
                     } else {
                         console.log('Posted an image!');
                     }
@@ -47,11 +39,12 @@ const upload_random_image = (images, statusText) => {
     });
 };
 
-
+// BOT PART
 fs.readdir(__dirname + '/images', (err, files) => {
     if (err) {
         console.log(err);
     } else {
+        // Pushing filenames into a new array to contain them all
         const images = [];
         files.forEach(f => {
             images.push(f);
@@ -59,26 +52,6 @@ fs.readdir(__dirname + '/images', (err, files) => {
         
         setInterval(() => {
             upload_random_image(images, 'Daily Yukine');
-        }, 82800000);
+        }, 7200000);
     }
 });
-
-
-// ! User stream is deprecated, feature on hold until I find another way
-/*
-const replyZenbu = msgContent => {
-    const replyto = msgContent.in_reply_to_screen_name;
-    const text = msgContent.text.toLowerCase();
-    const from = msgContent.user.screen_name;
-
-    console.log(replyto + ' ' + from);
-
-    if (replyto === 'yukinebotine' && text.includes('post chris')) {
-        const statusText = `@${from} Zenbu!`;
-        upload_random_image(imagesArray, statusText);
-    }
-}
-
-const stream = T.stream('user');
-stream.on('tweet', replyZenbu);
-*/
